# design/api/openapi.yml
openapi: 3.0.3
info:
  title: Point of Service (POS) API
  version: '1.0'
  description: 'REST API to be used to create tableside food ordering applications.'
  contact:
    email: mawentowski@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:4010/
    description: Prism proxy server
  - url: http://localhost:8080/
    description: API server
...
paths:
...
  /menus:
    post:
      summary: Create a menu
      description: This operation creates a new menu with the provided information.
      operationId: postMenu
      tags:
        - Menus
      security:
        - AccessToken: []
      requestBody:
        $ref: '#/components/requestBodies/CreateMenuRequest'
      responses:
        '201':
          description: Resource created successfully. The response body contains the newly created resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Menu'
          headers:
            Location:
              schema:
                type: string
              description: Points to the resource endpoint for the newly created resource
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description: Unauthorized. Authentication is required, or the provided credentials are invalid.
        '403':
          description: Forbidden. The server understood the request, but it refuses to authorize it. Authentication will not help.
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: The request could not be completed due to a conflict with the current state of the resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error. An unexpected error occurred on the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    get:
      summary: Get a list of menus
      description: This operation retrieves a list of menus based on the specified criteria.
      operationId: getMenuList
      tags:
        - Menus
      parameters:
        - name: sort
          in: query
          description: |
            Sort fields in ascending or descending order. To sort ascending, postfix the field name with `:asc`. To sort descending, postfix the field name with `:desc`.
          schema:
            type: string
          example: 'name:asc'
        - name: include
          in: query
          description: |
            Projection filters for including only certain resource properties in the response. The value for the parameter should be a comma-separated list of properties from the schema.
          schema:
            type: string
          example: 'name,category_ids'
        - name: select
          in: query
          description: |
            Selection filters to display resources matching the provided criteria. Use the format `fieldName.operator:value`.

            **Basic Equality (eq):**
            - For exact matches, use `eq:value`.

            **Comparison Operators:**
            - **Less Than (lt):** Use `lt:value`.
            - **Less Than or Equal To (lte):** Use `lte:value`.
            - **Greater Than (gt):** Use `gt:value`.
            - **Greater Than or Equal To (gte):** Use `gte:value`.

            Separate multiple filters with commas. Dates should be in `YYYY-MM-DD` format.
          schema:
            type: string
            example: 'created_at.gt:2024-04-23,name.eq:dinner'
        - name: limit
          in: query
          description: |
            Pagination filter for the maximum number of results to return in the response. If `limit` is used in conjunction with `offset`, then page-based pagination is applied.
          schema:
            type: integer
            maximum: 30
          example: 20
        - name: offset
          in: query
          description: |
            Offest-based pagination filter for the number of results to skip before displaying resources. A value of `0` will display resources from the beginning of the list. Only affects the response data if a `limit` value is set.
          schema:
            type: integer
            example: 10
      responses:
        '200':
          description: Successful operation. The response body contains the requested data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Menu'
                  total_results:
                    type: integer
                    example: 1
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description: Unauthorized. Authentication is required, or the provided credentials are invalid.
        '403':
          description: Forbidden. The server understood the request, but it refuses to authorize it. Authentication will not help.
        '500':
          description: Internal server error. An unexpected error occurred on the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /menus/{menu_id}:
    get:
      summary: View a menu
      description: This operation retrieves information about a menu identified by its unique menu ID.
      operationId: getMenu
      tags:
        - Menus
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
          example: '6220a6a8d01d24001f27d3ff'
          description: The unique identifier for the resource.
        - name: include
          in: query
          description: |
            Projection filters for including only certain resource properties in the response. The value for the parameter should be a comma-separated list of properties from the schema.
          schema:
            type: string
          example: 'name,category_ids'
      responses:
        '200':
          description: Successful operation. The response body contains the requested data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Menu'
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description: Unauthorized. Authentication is required, or the provided credentials are invalid.
        '403':
          description: Forbidden. The server understood the request, but it refuses to authorize it. Authentication will not help.
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error. An unexpected error occurred on the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    put:
      summary: Update a menu
      description: This operation updates an existing menu with the provided data.
      operationId: putMenu
      tags:
        - Menus
      security:
        - AccessToken: []
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
          example: '62a26aa41c7c5c001f627a4f'
          description: The unique identifier for the resource.
      requestBody:
        $ref: '#/components/requestBodies/UpdateMenuRequest'
      responses:
        '200':
          description: Successful operation. The response body contains the updated resource.
        '204':
          description: No content. The operation was successful, and there is no content to return.
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description: Unauthorized. Authentication is required, or the provided credentials are invalid.
        '403':
          description: Forbidden. The server understood the request, but it refuses to authorize it. Authentication will not help.
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: The request could not be completed due to a conflict with the current state of the resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error. An unexpected error occurred on the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    delete:
      summary: Delete a menu
      description: This operation deletes an existing menu identified by its unique menu ID.
      operationId: deleteMenu
      tags:
        - Menus
      security:
        - AccessToken: []
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
          example: '6261c0e33e5cf4f29b4641c4'
          description: The unique identifier for the resource.
      responses:
        '204':
          description: No content. The operation was successful, and there is no content to return.
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description: Unauthorized. Authentication is required, or the provided credentials are invalid.
        '403':
          description: Forbidden. The server understood the request, but it refuses to authorize it. Authentication will not help.
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error. An unexpected error occurred on the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /menus/{menu_id}/categories:
    get:
      summary: View a menu's categories
      description: This operation retrieves a list of categories belonging to a specific menu.
      operationId: getMenuCategories
      tags:
        - Menus
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
          example: '62f7c50e47410a001f512d4d'
          description: The unique identifier for the resource.
        - name: include
          in: query
          description: |
            Projection filters for including only certain resource properties in the response. The value for the parameter should be a comma-separated list of properties from the schema.
          schema:
            type: string
          example: 'name,dish_ids'
      responses:
        '200':
          description: Successful operation. The response body contains the requested data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description: Unauthorized. Authentication is required, or the provided credentials are invalid.
        '403':
          description: Forbidden. The server understood the request, but it refuses to authorize it. Authentication will not help.
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error. An unexpected error occurred on the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  requestBodies:
...
    CreateMenuRequest:
      description: Request body containing information used to create the resource.
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              category_ids:
                $ref: '#/components/schemas/CategoryIds'
              name:
                $ref: '#/components/schemas/Name'
            required:
              - name
              - category_ids
    UpdateMenuRequest:
      description: Request body containing information used to update the resource.
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              category_ids:
                $ref: '#/components/schemas/CategoryIds'
              name:
                $ref: '#/components/schemas/Name'
...
  schemas:
...
    Category:
      type: object
      properties:
        category_id:
          $ref: '#/components/schemas/Id'
        created_at:
          $ref: '#/components/schemas/CreatedAt'
        updated_at:
          $ref: '#/components/schemas/UpdatedAt'
        dish_ids:
          $ref: '#/components/schemas/DishIds'
        name:
          $ref: '#/components/schemas/Name'
    CategoryIds:
      type: array
      items:
        type: string
      example:
        - '66294b2a4475a41f3e709bbc'
        - '66294b2a4475a41f3e709bbd'
        - '66294b2a4475a41f3e709bbe'
      description: Contains the unique identifiers of categories to add to the menu.
    CreatedAt:
      type: string
      format: date-time
      example: '2024-04-22T10:00:00Z'
      readOnly: true
      description: Timestamp when the resource was created in ISO 8601 format.
...
    DishIds:
      type: array
      items:
        type: string
      example:
        - '692a4a34a46ecf001f35ac6f'
        - '6a4a857c022eb7001f9b8972'
        - '6b7c2441a0db43001f9bf377'
      description: Contains the unique identifiers of dishes to add to the order.
...
    Id:
      type: string
      example: '674d0bf5c28b69001f8e03a1'
      readOnly: true
      description: The unique identifier for the resource.
...
    Name:
      type: string
      example: 'main_course'
      description: ''
    Menu:
      type: object
      properties:
        order_id:
          $ref: '#/components/schemas/Id'
        created_at:
          $ref: '#/components/schemas/CreatedAt'
        updated_at:
          $ref: '#/components/schemas/UpdatedAt'
        category_ids:
          $ref: '#/components/schemas/CategoryIds'
        name:
          $ref: '#/components/schemas/Name'
...
    Problem:
      type: object
      properties:
        title:
          type: string
          description: Human-readable error title.
          example: Human-readable error title.
        detail:
          type: string
          description: Human-readable error details.
          example: Human-readable error details.
      required:
        - title
        - detail
...
    UpdatedAt:
      type: string
      format: date-time
      example: '2024-04-22T10:00:00Z'
      readOnly: true
      description: Timestamp when the order was created in ISO 8601 format.
...
  securitySchemes:
    AccessToken:
      type: http
      scheme: bearer
    BasicAuth:
      type: http
      scheme: basic
      description: |
        Authentication using client credentials. Use the `client_id` as the username and `client_secret` as the password.

        - **client_id**: Public identifier for the client application. Example: `admin_web_app`.
        - **client_secret**: Confidential string for client authentication. Example: `a5d7f23e-8b64-4b4c-9b11-21c5cfdf25f1`.